refactor(core): Introduce `ManualExecutionService`
