+++ packages/cli/src/workflow-runner.ts
@@ -2,14 +2,7 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-shadow */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import * as a from 'assert/strict';
-import {
-	DirectedGraph,
-	InstanceSettings,
-	WorkflowExecute,
-	filterDisabledNodes,
-	recreateNodeExecutionStack,
-} from 'n8n-core';
+import { InstanceSettings, WorkflowExecute } from 'n8n-core';
 import type {
 	ExecutionError,
 	IDeferredPromise,
@@ -19,8 +12,6 @@ import type {
 	WorkflowExecuteMode,
 	WorkflowHooks,
 	IWorkflowExecutionDataProcess,
-	IRunExecutionData,
-	IWorkflowExecuteAdditionalData,
 } from 'n8n-workflow';
 import {
 	ErrorReporterProxy as ErrorReporter,
@@ -40,12 +31,12 @@ import type { ScalingService } from '@/scaling/scaling.service';
 import type { Job, JobData } from '@/scaling/scaling.types';
 import { PermissionChecker } from '@/user-management/permission-checker';
 import * as WorkflowExecuteAdditionalData from '@/workflow-execute-additional-data';
-import * as WorkflowHelpers from '@/workflow-helpers';
 import { generateFailedExecutionFromError } from '@/workflow-helpers';
 import { WorkflowStaticDataService } from '@/workflows/workflow-static-data.service';
 
 import { ExecutionNotFoundError } from './errors/execution-not-found-error';
 import { EventService } from './events/event.service';
+import { ManualExecutionService } from './manual-execution.service';
 
 @Service()
 export class WorkflowRunner {
@@ -63,6 +54,7 @@ export class WorkflowRunner {
 		private readonly permissionChecker: PermissionChecker,
 		private readonly eventService: EventService,
 		private readonly instanceSettings: InstanceSettings,
+		private readonly manualExecutionService: ManualExecutionService,
 	) {}
 
 	/** The process did error */
@@ -297,7 +289,13 @@ export class WorkflowRunner {
 				);
 				workflowExecution = workflowExecute.processRunExecutionData(workflow);
 			} else {
-				workflowExecution = this.runManually(data, workflow, additionalData, executionId, pinData);
+				workflowExecution = this.manualExecutionService.runManually(
+					data,
+					workflow,
+					additionalData,
+					executionId,
+					pinData,
+				);
 			}
 
 			this.activeExecutions.attachWorkflowExecution(executionId, workflowExecution);
@@ -459,93 +457,5 @@ export class WorkflowRunner {
 		});
 
 		this.activeExecutions.attachWorkflowExecution(executionId, workflowExecution);
-	}
-
-	// eslint-disable-next-line @typescript-eslint/promise-function-async
-	runManually(
-		data: IWorkflowExecutionDataProcess,
-		workflow: Workflow,
-		additionalData: IWorkflowExecuteAdditionalData,
-		executionId: string,
-		pinData?: IPinData,
-	) {
-		if (data.triggerToStartFrom?.data && data.startNodes && !data.destinationNode) {
-			this.logger.debug(
-				`Execution ID ${executionId} had triggerToStartFrom. Starting from that trigger.`,
-				{ executionId },
-			);
-			const startNodes = data.startNodes.map((data) => {
-				const node = workflow.getNode(data.name);
-				a.ok(node, `Could not find a node named "${data.name}" in the workflow.`);
-				return node;
-			});
-			const runData = { [data.triggerToStartFrom.name]: [data.triggerToStartFrom.data] };
-
-			const { nodeExecutionStack, waitingExecution, waitingExecutionSource } =
-				recreateNodeExecutionStack(
-					filterDisabledNodes(DirectedGraph.fromWorkflow(workflow)),
-					new Set(startNodes),
-					runData,
-					data.pinData ?? {},
-				);
-			const executionData: IRunExecutionData = {
-				resultData: { runData, pinData },
-				executionData: {
-					contextData: {},
-					metadata: {},
-					nodeExecutionStack,
-					waitingExecution,
-					waitingExecutionSource,
-				},
-			};
-
-			const workflowExecute = new WorkflowExecute(additionalData, 'manual', executionData);
-			return workflowExecute.processRunExecutionData(workflow);
-		} else if (
-			data.runData === undefined ||
-			data.startNodes === undefined ||
-			data.startNodes.length === 0
-		) {
-			// Full Execution
-			// TODO: When the old partial execution logic is removed this block can
-			// be removed and the previous one can be merged into
-			// `workflowExecute.runPartialWorkflow2`.
-			// Partial executions then require either a destination node from which
-			// everything else can be derived, or a triggerToStartFrom with
-			// triggerData.
-			this.logger.debug(`Execution ID ${executionId} will run executing all nodes.`, {
-				executionId,
-			});
-			// Execute all nodes
-
-			const startNode = WorkflowHelpers.getExecutionStartNode(data, workflow);
-
-			// Can execute without webhook so go on
-			const workflowExecute = new WorkflowExecute(additionalData, data.executionMode);
-			return workflowExecute.run(workflow, startNode, data.destinationNode, data.pinData);
-		} else {
-			// Partial Execution
-			this.logger.debug(`Execution ID ${executionId} is a partial execution.`, { executionId });
-			// Execute only the nodes between start and destination nodes
-			const workflowExecute = new WorkflowExecute(additionalData, data.executionMode);
-
-			if (data.partialExecutionVersion === '1') {
-				return workflowExecute.runPartialWorkflow2(
-					workflow,
-					data.runData,
-					data.pinData,
-					data.dirtyNodeNames,
-					data.destinationNode,
-				);
-			} else {
-				return workflowExecute.runPartialWorkflow(
-					workflow,
-					data.runData,
-					data.startNodes,
-					data.destinationNode,
-					data.pinData,
-				);
-			}
-		}
 	}
 }
